postgres-pgvector-extension/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
comm==0.2.0
dataclasses-json==0.6.2
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
exceptiongroup==1.1.3
executing==2.0.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.5.0
langchain==0.0.336
langsmith==0.0.64
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.2
openai==1.3.2
packaging==23.2
parso==0.8.3
pexpect==4.8.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.3
tenacity==8.2.3
tiktoken==0.5.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.10
yarl==1.9.2
